Replaced moves-enemies to tick-enemy
Shifted map call to tick-handler, changed posn-changer so that it worked
on a posn and a velvect in order to deal with all changes to an Enemy
more comprehensibly. Added vel-changer to tick-enemy, which has a
constant chance of randomizing vel each time it is called. Changed unit
tests to account for updated function names, signatures, and
randomization.